FEATURE: Improve redirection logic (#92)
* FEATURE: Improve redirection logic

* resolve comments

* resolve comments
