+++ src/constants/router.ts
@@ -1,0 +1,3 @@
+export const publicRoutes = ['/profile/settings', '/profile/applications', '/fill-the-form']
+
+export const privateRoutes = ['/login', '/register', '/reset-password']
