+++ src/pages/_app.tsx
@@ -14,6 +14,8 @@ import Head from 'next/head'
 
 import '@utils/i18n'
 
+import Middleware from '../Middleware'
+
 import '@styles/globals.css'
 
 const roboto = Roboto({
@@ -83,15 +85,17 @@ const App = ({ Component, pageProps }: IAppProps) => {
       <main>
         <Provider store={store}>
           <RedirectionHandler />
-          <Layout>
-            {HierarchicalLayout ? (
-              <HierarchicalLayout>
+          <Middleware>
+            <Layout>
+              {HierarchicalLayout ? (
+                <HierarchicalLayout>
+                  <Component {...pageProps} />
+                </HierarchicalLayout>
+              ) : (
                 <Component {...pageProps} />
-              </HierarchicalLayout>
-            ) : (
-              <Component {...pageProps} />
-            )}
-          </Layout>
+              )}
+            </Layout>
+          </Middleware>
           <ModalsController />
         </Provider>
       </main>
